# 02-use-cases/01-conversational-agents/local-prototype-to-agentcore/local_prototype/local_insurance_api/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Auto Insurance API",
    "version": "0.1.0",
    "description": "API for auto insurance services including customer information, vehicle data, and policy management"
  },
  "paths": {
...
    "/policies": {
      "get": {
        "summary": "Get All Policies",
        "description": "Get all policies",
        "operationId": "getAllPolicies",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PoliciesResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Get Filtered Policies",
        "description": "Get policies with optional filtering\n\nOptional parameters:\n- policy_id: string - Filter by specific policy ID\n- customer_id: string - Filter by customer ID\n- status: string - Filter by status (active, expired, etc.)\n- include_vehicles: boolean - Include vehicle details in response (default: true)",
        "operationId": "filterPolicies",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FilteredPoliciesRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PoliciesResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/policies/{policy_id}": {
      "get": {
        "summary": "Get Policy By Id",
        "description": "Get a specific policy by ID",
        "operationId": "getPolicyById",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Policy Id",
              "type": "string"
            },
            "name": "policy_id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyResponse"
                }
              }
            }
          },
          "404": {
            "description": "Policy not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/customer/{customer_id}/policies": {
      "get": {
        "summary": "Get Customer Policies",
        "description": "Get all policies for a specific customer",
        "operationId": "getCustomerPolicies",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Customer Id",
              "type": "string"
            },
            "name": "customer_id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerPoliciesResponse"
                }
              }
            }
          },
          "404": {
            "description": "Customer not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "FilteredPoliciesRequest": {
        "title": "Filtered Policies Request",
        "type": "object",
        "properties": {
          "policy_id": {
            "type": "string",
            "example": "POL-12345"
          },
          "customer_id": {
            "type": "string",
            "example": "C12345"
          },
          "status": {
            "type": "string",
            "enum": ["active", "expired", "cancelled", "pending"],
            "example": "active"
          },
          "include_vehicles": {
            "type": "boolean",
            "default": true,
            "example": true
          }
        }
      },
      "PoliciesResponse": {
        "title": "Policies Response",
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "success"
          },
          "count": {
            "type": "integer",
            "example": 2
          },
          "policies": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Policy"
            }
          }
        }
      },
      "PolicyResponse": {
        "title": "Policy Response",
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "success"
          },
          "policy": {
            "$ref": "#/components/schemas/Policy"
          }
        }
      },
      "CustomerPoliciesResponse": {
        "title": "Customer Policies Response",
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "success"
          },
          "customer_id": {
            "type": "string",
            "example": "C12345"
          },
          "count": {
            "type": "integer",
            "example": 2
          },
          "policies": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Policy"
            }
          }
        }
      },
      "Policy": {
        "title": "Policy",
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "POL-12345"
          },
          "customer_id": {
            "type": "string",
            "example": "C12345"
          },
          "type": {
            "type": "string",
            "example": "auto"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "example": "2025-01-01"
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "example": "2025-07-01"
          },
          "premium": {
            "type": "number",
            "example": 750
          },
          "status": {
            "type": "string",
            "enum": ["active", "expired", "cancelled", "pending"],
            "example": "active"
          },
          "coverage": {
            "type": "object",
            "properties": {
              "liability": {
                "type": "number",
                "example": 100000
              },
              "collision": {
                "type": "boolean",
                "example": true
              },
              "comprehensive": {
                "type": "boolean",
                "example": true
              },
              "uninsured_motorist": {
                "type": "boolean",
                "example": true
              },
              "rental_reimbursement": {
                "type": "boolean",
                "example": false
              },
              "roadside_assistance": {
                "type": "boolean",
                "example": true
              }
            }
          },
          "vehicles": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "make": {
                  "type": "string",
                  "example": "Toyota"
                },
                "model": {
                  "type": "string",
                  "example": "Camry"
                },
                "year": {
                  "type": "string",
                  "example": "2023"
                },
                "vin": {
                  "type": "string",
                  "example": "1HGCM82633A123456"
                }
              }
            }
          }
        }
      },
      "ErrorResponse": {
        "title": "Error Response",
        "type": "object",
        "properties": {
          "detail": {
            "type": "string",
            "example": "Customer not found"
          }
        }
      },
      "HTTPValidationError": {
        "title": "HTTPValidationError",
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
      "ValidationError": {
        "title": "ValidationError",
        "required": ["loc", "msg", "type"],
        "type": "object",
        "properties": {
          "loc": {
            "title": "Location",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        }
      }
    }
  }
}